ref(normalization): Restore TransactionsProcessor
